requirements.txt
Flask==2.2.2
gunicorn==20.1.0
google-cloud-pubsub==2.13.6
google-cloud-secret-manager==2.12.4
marshmallow==3.18.0
CouchDB~=1.2
SQLAlchemy==1.4.41
pg8000==1.29.1